Separating property widgets in files
